+++ src/main/java/com/pradip/sewearn/dto/MaterialEarningDTO.java
@@ -7,5 +7,5 @@ import lombok.Data;
 @Builder
 public class MaterialEarningDTO {
     private String materialName;
-    private Double earning;
+    private Long earning;
 }
