+++ src/main/java/com/pradip/sewearn/model/submit/SewEarnSubmit.java
@@ -22,7 +22,7 @@ public class SewEarnSubmit {
 
     private LocalDate submissionDate;
 
-    private Double totalEarning; // sum of all submitted items
+    private Long totalEarning;
 
     private Integer totalQuantity;
